Seperated LoginForm into own file
