+++ client/src/components/LoginModal/LoginModalTemplate.jsx
@@ -1,44 +1,9 @@
 // USED AS TEMPLATE BY BOTH LOGIN AND SIGNUP MODALS
 import React from "react";
 
-import { useFormik } from 'formik';
-
 import crown from "../../assets/crown.png";
+import LoginForm from "./LoginForm";
 import "./LoginModal.css";
-
-const Form = (props) => {
-  const { submitButtonText, submitButtonAction } = props;
-
-  const formik = useFormik({
-    initialValues: {
-      email: '',
-      password: '',
-    },
-    onSubmit: values => submitButtonAction(values)
-  });
-  return (
-    <form onSubmit={formik.handleSubmit} className="loginModal">
-      <label htmlFor="email">Email Address</label>
-      <input
-        id="email"
-        name="email"
-        type="email"
-        onChange={formik.handleChange}
-        value={formik.values.email}
-      />
-
-      <label htmlFor="password">Password:</label>
-      <input
-        type="password"
-        name="password"
-        onChange={formik.handleChange}
-        value={formik.values.password}
-      />
-
-      <button type="submit">{submitButtonText}</button>
-    </form>
-  );
-};
   
 const LoginModalTemplate = (props) => {
   const { headerImage, showModal, hideModal, ...rest } = props;
@@ -61,7 +26,7 @@ const LoginModalTemplate = (props) => {
       onClick={(e) => closeModal(e)} // close modal if user clicks outside of modal
     >
       {header}
-      <Form {...rest} />
+      <LoginForm {...rest} />
     </div>
   );
 };
